requirements.txt
# Icons8 Collector Dependencies
# Install with: pip install -r requirements.txt
# Then install Playwright browsers: python -m playwright install chromium

requests>=2.28.0,<3.0.0
pillow>=9.0.0
playwright>=1.40.0,<2.0.0
tenacity>=8.0.0,<9.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-mock>=3.10.0
# responses>=0.23.0
